Add files via upload
Tf-idf stands for “Term Frequency - Inverse Document Frequency”. It determines the importance of words in a set of documents. It is the product of Term Frequency and
Inverse Document Frequency. Term Frequency is used to measure the frequency of words in a document. 

In this Assignment, I call tf-idf code to convert 200 reviews or documents into term by document matrix.

Singular value decompostion is applied to reduce the number of terms or features to 100.

Logistic Regression and Neural Networks are applied to the data to obtain 5-cross validation errors.